MODIFIED Cash register viewmodel and view controller to be MVVM
